Summary improvements (#8)
Improve summary prompts and use logger. Also auto-correct several style issues.

+++ worklog/storage.rb
@@ -13,7 +13,7 @@ module Storage
 
   # Return all days with logs
   def self.all_days
-    return [] if !folder_exists?
+    return [] unless folder_exists?
 
     logs = []
     Dir.glob(File.join(DATA_DIR, "*#{FILE_SUFFIX}")).map do |file|
@@ -26,7 +26,7 @@ module Storage
   # Return days between start_date and end_date
   # If end_date is nil, return logs from start_date to today
   def self.days_between(start_date, end_date = nil, epics_only = nil, tags_filter = nil)
-    return [] if !folder_exists?
+    return [] unless folder_exists?
 
     logs = []
     end_date = Date.today if end_date.nil?
@@ -43,9 +43,7 @@ module Storage
           tmp_logs.entries.keep_if { |entry| entry.tags && (entry.tags & tags_filter).size > 0 }
         end
 
-        if tmp_logs.entries.length > 0
-          logs << tmp_logs
-        end
+        logs << tmp_logs if tmp_logs.entries.length > 0
       end
 
       start_date += 1
@@ -57,18 +55,14 @@ module Storage
   def self.create_file_skeleton(date)
     create_folder
 
-    unless File.exist?(filepath(date))
-      File.write(filepath(date), YAML.dump(DailyLog.new(date, [])))
-    end
+    File.write(filepath(date), YAML.dump(DailyLog.new(date, []))) unless File.exist?(filepath(date))
   end
 
   def self.load_log(file)
     WorkLogger.debug "Loading file #{file}"
     log = YAML.load_file(file, permitted_classes: [Date, Time, DailyLog, LogEntry])
     log.entries.each do |entry|
-      unless entry.time.respond_to?(:strftime)
-        entry.time = Time.parse(entry.time)
-      end
+      entry.time = Time.parse(entry.time) unless entry.time.respond_to?(:strftime)
     end
     log
   end
@@ -85,9 +79,7 @@ module Storage
 
   def self.load_single_log_file(file, headline = true)
     daily_log = load_log(file)
-    if headline
-      puts "Work log for #{Rainbow(daily_log.date).gold}:"
-    end
+    puts "Work log for #{Rainbow(daily_log.date).gold}:" if headline
     daily_log.entries
   end
 
@@ -95,9 +87,7 @@ module Storage
 
   # Create folder if not exists already.
   def create_folder
-    unless Dir.exist?(DATA_DIR)
-      Dir.mkdir(DATA_DIR)
-    end
+    Dir.mkdir(DATA_DIR) unless Dir.exist?(DATA_DIR)
   end
 
   def filepath(date)
